requirements.txt
# Deal Logging Bot Dependencies
python-telegram-bot>=20.7
notion-client>=2.2.1
openai>=1.12.0  # Used for Kimi API (OpenAI-compatible)
python-dotenv>=1.0.0
playwright>=1.41.0
img2pdf>=0.5.0
Pillow>=10.0.0
pypdf>=3.17.0
aiohttp>=3.9.0
